\name{getCorrection}
\alias{getCorrection}
\title{Correction} 
\description{
  This function computes the corrected intensity.
}
\usage{
getCorrection(effet, model, regTab)
}
\arguments{
  \item{effet}{The name of the biological effect}
  \item{model}{The class lm object given by the \code{getModel} function}
  \item{regTab}{The regression table used to estimate the linear model, and containing the variables in the model}
}
\note{People interested in tools dealing with array CGH analysis and DNA copy number analysis can
  visit our web-page \url{http://bioinfo.curie.fr}.}

\author{Guillem Rigaill, \email{italics@curie.fr}.}
\source{Institut Curie, \email{italics@curie.fr}.}
\keyword{}

