\name{getModel}
\alias{getModel}
\title{Regression Model} 
\description{
 Computes the linear regression model and and return an object of class lm.
}
\usage{
getModel(formule, response, regTab)
}
\arguments{
  \item{formule}{A symbolic description of the term of the model. It is a string}
  \item{response}{The parameter you want to explain (the response) : the SNP "LogRatio". Y is a string}
  \item{regTab}{A Regression table containing the variables in the model}
}
\note{People interested in tools dealing with array CGH analysis and DNA copy number analysis can
  visit our web-page \url{http://bioinfo.curie.fr}.}


\author{Guillem Rigaill, \email{italics@curie.fr}.}
\source{Institut Curie, \email{italics@curie.fr}.}
\keyword{}

