\name{getResidu}
\alias{getResidu}
\title{Correction} 
\description{
  This function retrieves the residual values
}
\usage{
getResidu(model)
}
\arguments{
  \item{model}{The class lm object given by the \code{getModel} function}
}
\note{People interested in tools dealing with array CGH analysis and DNA copy number analysis can
  visit our web-page \url{http://bioinfo.curie.fr}.}

\author{Guillem Rigaill, \email{italics@curie.fr}.}
\source{Institut Curie, \email{italics@curie.fr}.}
\keyword{}

